+++ mopidy/config/__init__.py
@@ -177,13 +177,19 @@ def _validate(raw_config, schemas):
     # Get validated config
     config = {}
     errors = {}
+    sections = set(raw_config)
     for schema in schemas:
+        sections.discard(schema.name)
         values = raw_config.get(schema.name, {})
         result, error = schema.deserialize(values)
         if error:
             errors[schema.name] = error
         if result:
             config[schema.name] = result
+
+    for section in sections:
+        logger.debug('Ignoring unknown config section: %s', section)
+
     return config, errors